Bug/mes 295

+++ src/main/java/com/alcegory/mescloud/model/dto/CountingEquipmentDto.java
@@ -21,7 +21,7 @@ public class CountingEquipmentDto {
     private List<EquipmentOutputDto> outputs;
     private ImsDto ims;
 
-    private Integer theoreticalProduction;
+    private Double theoreticalProduction;
     private Double qualityTarget;
     private Double availabilityTarget;
     private Double performanceTarget;
